Extract attachable endpoint from attachment api
