+++ app/api/api.rb
@@ -176,6 +176,7 @@ class API < Grape::API
   mount Chemotion::NmrdbAPI
   mount Chemotion::MeasurementsAPI
   mount Chemotion::ConverterAPI
+  mount Chemotion::AttachableAPI
 
   add_swagger_documentation(info: {
     "title": "Chemotion ELN",
