+++ spec/api/chemotion/attachable_api_spec.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+describe Chemotion::AttachableAPI do
+  describe 'POST /api/v1/attachable/update_attachments_attachable' do
+    # TODO: Add a spec for this endpoint
+    pending 'not yet implemented'
+  end
+end
